fix(ssh): run exec-channel commands through sh, not the login shell
sshd runs an exec-channel command as `$SHELL -c '<command>'`, so the
remote user's *login* shell parses it. `TrzszConfig.serverCommand()` and
`MoshConfig.serverCommand()` handed it the bare POSIX-sh PATH prelude,
which fish rejects at the first token — `Unsupported use of '='` — so
tsshd and mosh-server never started on a host whose login shell is fish
or csh. Every other consumer of the prelude already wraps its payload in
`sh -c`/`sh -lc`; these two did not.

Both now wrap through `RemoteLoginShell.wrapForLoginShell`, a new
dependency-free home for the prelude and its quoting so a standalone
package can symlink it and syntax-check what remote shells actually
parse. `exec` before the server binary keeps the process tree the same
depth as before, so the SSH channel's signals and EOF still reach the
server directly.

The quoting idiom changes with it. `'\''` relies on the outer shell
reading `\'` inside a single-quoted string as a literal backslash then
quote; fish instead reads it as an escaped quote. The two agree at one
level of nesting and diverge at two, which mosh reaches for real when a
TERM override wraps a zmx/tmux/herdr attach line inside the new wrap.
`'"'"'` carries no backslash for the two grammars to disagree about and
nests to arbitrary depth in sh, bash, zsh and fish alike.
